# paths and simd perf improvements