--- a/experiments/benchmark_svg_cairo.nim
+++ b/experiments/benchmark_svg_cairo.nim
@@ -1,6 +1,0 @@
-import benchy, svg_cairo
-
-let data = readFile("tests/fileformats/svg/Ghostscript_Tiger.svg")
-
-timeIt "svg decode":
-  discard decodeSvg(data)