… experiments/svg_cairo.nim
… src/pixie/internal.nim
… src/pixie/masks.nim
… src/pixie/paths.nim
… tests/masks/drawPolygonMagnified.png
… tests/test_masks.nim