before:
```
rect Image NormalBlend ............. 0.048 ms      0.050 ms    ±0.003  x1000
rect Image SubtractMaskBlend ....... 0.473 ms      0.487 ms    ±0.007  x1000
rect Image ExcludeMaskBlend ........ 0.331 ms      0.340 ms    ±0.007  x1000
roundedRect Image OverwriteBlend ... 0.070 ms      0.076 ms    ±0.012  x1000
roundedRect Image NormalBlend ...... 0.134 ms      0.143 ms    ±0.020  x1000
roundedRect Image SubtractMaskBlend  0.535 ms      0.552 ms    ±0.017  x1000
roundedRect Image ExcludeMaskBlend . 0.433 ms      0.437 ms    ±0.005  x1000
```

after:
```
rect Image NormalBlend ............. 0.033 ms      0.037 ms    ±0.008  x1000
rect Image SubtractMaskBlend ....... 0.374 ms      0.377 ms    ±0.006  x1000
rect Image ExcludeMaskBlend ........ 0.269 ms      0.272 ms    ±0.004  x1000
roundedRect Image OverwriteBlend ... 0.060 ms      0.061 ms    ±0.001  x1000
roundedRect Image NormalBlend ...... 0.132 ms      0.135 ms    ±0.007  x1000
roundedRect Image SubtractMaskBlend  0.455 ms      0.463 ms    ±0.022  x1000
roundedRect Image ExcludeMaskBlend . 0.382 ms      0.385 ms    ±0.004  x1000
```

----

before:
```
[pixie] rect ....................... 0.075 ms      0.092 ms    ±0.014  x1000
[pixie] roundedRect ................ 0.111 ms      0.113 ms    ±0.002  x1000
[pixie] heart ...................... 0.279 ms      0.284 ms    ±0.008  x1000
```

after:
```
[pixie] rect ....................... 0.074 ms      0.087 ms    ±0.013  x1000
[pixie] roundedRect ................ 0.103 ms      0.105 ms    ±0.003  x1000
[pixie] heart ...................... 0.244 ms      0.247 ms    ±0.005  x1000
```

----

before:
```
svg render ........................ 11.761 ms     11.832 ms    ±0.103   x422
```

after:
```
svg render ........................ 11.668 ms     11.752 ms    ±0.108   x425
```

----

before:
```
minifyBy2 .......................... 0.468 ms      0.471 ms    ±0.007  x1000 [mask]
minifyBy2 .......................... 1.180 ms      1.205 ms    ±0.031  x1000 [image]
isOneColor ......................... 0.306 ms      0.318 ms    ±0.024  x1000
magnifyBy2 ......................... 5.879 ms      6.437 ms    ±0.481   x766
invert ............................. 1.461 ms      1.480 ms    ±0.016  x1000
newMask(image) ..................... 0.529 ms      0.539 ms    ±0.016  x1000
newImage(mask) ..................... 0.831 ms      0.846 ms    ±0.020  x1000
toPremultipliedAlpha ............... 1.129 ms      1.140 ms    ±0.016  x1000
```

after:
```
minifyBy2 .......................... 0.277 ms      0.286 ms    ±0.021  x1000 [mask]
minifyBy2 .......................... 0.980 ms      1.002 ms    ±0.040  x1000 [image]
isOneColor ......................... 0.182 ms      0.207 ms    ±0.035  x1000
magnifyBy2 ......................... 5.846 ms      6.145 ms    ±0.350   x803
invert ............................. 1.383 ms      1.405 ms    ±0.030  x1000
newMask(image) ..................... 0.431 ms      0.441 ms    ±0.016  x1000
newImage(mask) ..................... 0.798 ms      0.814 ms    ±0.023  x1000
toPremultipliedAlpha ............... 1.023 ms      1.036 ms    ±0.026  x1000
```